wowzas 2

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/command/core/LiftCommands.kt
@@ -7,4 +7,4 @@ import org.firstinspires.ftc.teamcode.utility.DelayedCommand
 class LiftTo(target: Int, lift: Lift, delay: Long = 0) : DelayedCommand({ lift.target = target }, delay) {
     constructor(state: Lift.State, lift: Lift, delay: Long = 0): this(state.target, lift,  delay)
 }
-class NudgeLift(nudgeBy: Int, lift: Lift, delay: Long = 0): DelayedCommand({ lift.target = lift.position + nudgeBy }, delay)
+class NudgeLift(nudgeBy: Int, lift: Lift, delay: Long = 0): DelayedCommand({ lift.target = (lift.position + nudgeBy).toInt() }, delay)
