+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/hardware/subsystems/IExtendable.kt
@@ -1,6 +1,6 @@
 package org.firstinspires.ftc.teamcode.hardware.subsystems
 
 interface IExtendable : ISubsystem {
-    var position: Int
+    var position: Double
     var target: Int
 }
